refactor(web): redesign config pages and extract raw JSON editor
- reorganize visual config into card-based setting sections
- move the raw config editor into a dedicated full-page view
- simplify shared form layouts and remove unused locale strings
