+++ web/frontend/src/i18n/locales/en.json
@@ -331,33 +331,10 @@
   "pages": {
     "agent": {
       "load_error": "Failed to load agent support information.",
-      "stats": {
-        "workspace": "Workspace",
-        "workspace_hint": "The default agent workspace used for runtime files and workspace skills.",
-        "skills": "Available Skills",
-        "skills_hint": "Skills discovered from workspace, global, and builtin roots.",
-        "tools": "Enabled Tools",
-        "tools_hint": "{{blocked}} blocked by missing dependencies."
-      },
       "skills": {
-        "title": "Skills",
         "description": "Skills are loaded from the workspace, global PicoClaw home, and builtin directories.",
-        "hero_title": "Skill Library",
-        "hero_description": "Browse every capability package the agent can load, then drill straight into the effective SKILL.md without leaving the page.",
-        "stats": {
-          "total": "Total Skills",
-          "workspace": "Workspace",
-          "shared": "Shared"
-        },
         "empty": "No skills are currently available.",
         "import": "Import Skill",
-        "import_title": "Import Skill",
-        "import_description": "Create a workspace skill by uploading a markdown file as the new SKILL.md.",
-        "import_name": "Skill Name",
-        "import_name_placeholder": "e.g. my-workflow",
-        "import_file": "Markdown File",
-        "import_file_hint": "Upload a .md file. The backend stores it as workspace/skills/<name>/SKILL.md.",
-        "import_confirm": "Import Skill",
         "import_success": "Skill imported.",
         "import_error": "Failed to import skill.",
         "view": "View",
@@ -371,28 +348,11 @@
         "viewer_description": "Read the current effective SKILL.md content here.",
         "loading_detail": "Loading skill content...",
         "load_detail_error": "Failed to load skill content.",
-        "source": "Source",
         "path": "Skill Path",
-        "no_description": "No description provided.",
-        "sources": {
-          "workspace": "Workspace",
-          "global": "Global",
-          "builtin": "Builtin"
-        },
-        "errors": {
-          "file_required": "Please choose a markdown file to import."
-        }
+        "no_description": "No description provided."
       },
       "tools": {
-        "title": "Tools",
         "description": "This view reflects whether each agent tool is enabled, disabled, or blocked by a missing prerequisite.",
-        "hero_title": "Tool Surface",
-        "hero_description": "Inspect what the agent can actually call right now, which capabilities are blocked, and where each tool is controlled in config.",
-        "stats": {
-          "enabled": "Enabled",
-          "blocked": "Blocked",
-          "categories": "Categories"
-        },
         "empty": "No tools are available.",
         "enable": "Enable",
         "disable": "Disable",
@@ -468,13 +428,15 @@
       "allowed_cidrs": "Allowed Network CIDRs",
       "allowed_cidrs_hint": "Only clients from these CIDR ranges can access the service. One per line or comma-separated. Leave empty to allow all.",
       "allowed_cidrs_placeholder": "192.168.1.0/24\n10.0.0.0/8",
-      "launcher_load_error": "Failed to load service parameters.",
-      "launcher_restart_hint": "Service parameter changes apply after restarting PicoClaw Web.",
-      "advanced_desc": "Open the raw JSON page to edit every field directly.",
+      "sections": {
+        "agent": "Agent",
+        "runtime": "Runtime",
+        "launcher": "Service",
+        "devices": "Devices"
+      },
       "open_raw": "Raw Config",
       "back_to_visual": "Visual Config",
       "raw_json_title": "Raw JSON Configuration",
-      "raw_json_desc": "Advanced users can directly edit the raw JSON configuration below.",
       "json_placeholder": "Enter valid JSON configuration...",
       "save_success": "Configuration saved successfully.",
       "save_error": "Failed to save configuration.",
